feat: require at least admin access to create, delete and update books.

+++ src/Api/Controllers/BooksController.cs
@@ -1,6 +1,8 @@
 using Api.Adapters;
 using Api.DTOs.Request;
+using Api.Extensions;
 using Core.Services.Interfaces;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Api.Controllers;
@@ -33,14 +35,21 @@ public class BooksController : ControllerBase
         return Ok(response);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<IActionResult> Create([FromBody] CreateBookDTO bookDTO, CancellationToken cancellationToken)
     {
+        if (!User.HasAdminAccess())
+        {
+            return Forbid();
+        }
+
         var book = BookAdapter.ToDomainEntity(bookDTO);
         var createdBookId = await _bookService.CreateAsync(book, cancellationToken);
         return Ok(createdBookId);
     }
 
+    [Authorize]
     [HttpPatch]
     [Route("{bookId}")]
     public async Task<IActionResult> Update(
@@ -48,11 +57,17 @@ public class BooksController : ControllerBase
         [FromBody] UpdateBookDTO bookDTO,
         CancellationToken cancellationToken)
     {
+        if (!User.HasAdminAccess())
+        {
+            return Forbid();
+        }
+
         var book = BookAdapter.ToDomainEntity(bookDTO);
         await _bookService.UpdateAsync(bookId, book, cancellationToken);
         return NoContent();
     }
 
+    [Authorize]
     [HttpPatch]
     [Route("{bookId}/quantity")]
     public async Task<IActionResult> UpdateBookQuantity(
@@ -60,10 +75,16 @@ public class BooksController : ControllerBase
         [FromBody] UpdateBookQuantityDTO bookDTO,
         CancellationToken cancellationToken)
     {
+        if (!User.HasAdminAccess())
+        {
+            return Forbid();
+        }
+
         await _bookService.UpdateQuantityAsync(bookId, (int)bookDTO.Quantity!, cancellationToken);
         return NoContent();
     }
 
+    [Authorize]
     [HttpPut]
     [Route("{bookId}")]
     public async Task<IActionResult> Replace(
@@ -71,17 +92,28 @@ public class BooksController : ControllerBase
         [FromBody] ReplaceBookDTO bookDTO,
         CancellationToken cancellationToken)
     {
+        if (!User.HasAdminAccess())
+        {
+            return Forbid();
+        }
+
         var book = BookAdapter.ToDomainEntity(bookDTO);
         await _bookService.UpdateAsync(bookId, book, cancellationToken);
         return NoContent();
     }
 
+    [Authorize]
     [HttpDelete]
     [Route("{bookId}")]
     public async Task<IActionResult> Delete(
         [FromRoute] string bookId,
         CancellationToken cancellationToken)
     {
+        if (!User.HasAdminAccess())
+        {
+            return Forbid();
+        }
+
         await _bookService.DeleteAsync(bookId, cancellationToken);
         return NoContent();
     }
